refactor: reorganise frontend into feature folders
- features/solver/  — grid components, reducer, hooks, tests
- features/results/ — solver output display (ResultPanel)
- shared/ui/        — cross-cutting widgets (ErrorBanner, NotificationStack, DebugPane)

No behaviour change. All 33 tests pass.
