+++ frontend/src/features/results/components/ResultPanel.jsx
