+++ frontend/src/features/solver/__tests__/SolverShell.test.jsx
@@ -2,7 +2,7 @@ import React from 'react'
 import { describe, it, expect, vi, beforeEach } from 'vitest'
 import { render, screen, waitFor } from '@testing-library/react'
 import userEvent from '@testing-library/user-event'
-import { SolverShell } from '@/components/SolverShell'
+import { SolverShell } from '@/features/solver/components/SolverShell'
 
 // Mock the API client — tests must not make real network calls
 vi.mock('@/api/client', () => ({
