+++ frontend/src/features/solver/__tests__/solverReducer.test.js
@@ -7,7 +7,7 @@ import {
   cycleFeedback,
   isDraftValid,
   historyToSteps,
-} from '@/hooks/solverReducer'
+} from '../hooks/solverReducer'
 
 // ── cycleFeedback ──────────────────────────────────────────────────────────────
