+++ frontend/src/features/solver/components/GuessTile.jsx
