+++ frontend/src/features/solver/hooks/useConfetti.js
